requirements_pip.txt
efficientnet-pytorch>=0.7.0
imbalanced-learn>=0.7.0
imblearn>=0.0
matplotlib>=3.3.4
pandas>=1.2.2
regex>=2020.11.13
scikit-learn>=0.24.1
scipy>=1.6.0
torch-summary>=1.4.5
